requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
markdown2==2.4.10
python-dotenv==1.0.1